Replaced mount-manager with well tested k8s mount utilities
